Implement problem solution.
Implement a simple, lightweight solution of the problem using hashsets and minimal code.
It uses a Hashset to store visited houses, and (i32, i32) for positions.

Part 1 and Part 2 solutions are separated by different functions in the same module.

This solution is not compatible with the provided unit tests, because its structure is lacking the framework.
On the other hand, the advent of code website gave me two gold stars for the answer.